Create WebClient to talk to KimonoLabs' API.
WebClient is a lib class that talks to KimonoLabs to get available APIs or fetch an API's json.

Includes tests for the class and the list method.

Uses faraday gem for http connections and VCR gem to record those web api calls for replay
in test environments (e.g. Travis).