Configure VCR to work properly in test environments.
Change the default VCR config to match on paths and not full urls,
so that we can pass nonsense api keys and still playback the VCR cassette.

Add rake task to rerecord instead of using the :new_episodes option, to
guarantee we don't make live web request attempts in test.